backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-docx
docx2txt
PyMuPDF
fuzzywuzzy
python-Levenshtein
rank_bm25
langchain
langchain-google-generative-ai
sentence-transformers
transformers
tensorflow
tqdm
requests
numpy









# # Core Backend
# Flask==2.3.3
# Flask-CORS==4.0.0
# Flask-SQLAlchemy==3.0.5
# psycopg2-binary==2.9.7
# python-dotenv==1.0.0
# Werkzeug==2.3.7

# # File Parsing
# PyMuPDF
# python-docx==0.8.11
# docx2txt==0.8

# # NLP & Preprocessing
# spacy==3.6.1
# nltk==3.8.1
# scikit-learn==1.3.0
# rank-bm25==0.2.2
# fuzzywuzzy==0.18.0
# python-Levenshtein==0.21.1

# # LLM & Semantic Layer
# langchain==0.0.330
# langchain-google-genai==0.0.5
# langsmith==0.0.64
# sentence-transformers==2.2.2
# transformers==4.34.0
# torch==2.0.1

# # Vector Store
# chromadb==0.4.15
# faiss-cpu==1.7.4

# # Additional utilities
# pandas==2.1.1
# numpy==1.24.3
# python-dateutil==2.8.2
# requests==2.31.0

# # Testing
# pytest==7.4.2
# pytest-flask==1.2.0